- Added support for internal texture formats - Added a default frame buffer to camera render textures for storing object IDs - ModelViewProjection is now a more general object data struct - Implemented mouse picking - Added a way to switch between attachments of the scene cameras framebuffer for debugging purposes - Added support for a secondary readonly assets folder - Added a way to import default assets into a new project

+++ GloryEngine/GloryCore/ThreadManager.cpp
@@ -43,7 +43,7 @@ namespace Glory
 	{
 		//if (m_pThreads.size() >= m_MaxThreads) return false;
 
-		int index = m_pThreads.size();
+		size_t index = m_pThreads.size();
 
 		std::function<void(Thread*)> func = std::bind(&ThreadManager::OnThreadIdle, this, std::placeholders::_1);
 		Thread* newThread = new Thread(index, func);
